Add files via upload
This is my vignere cypher code. The encoder is two letters off to the right (so when the letter should be m, it is o instead). Don't know why that is. Also, caps lock doesn't work. And -e and -d implmentation isn't in there. Simply a rough prototype to be built on, but I think it's a good starting place.